+++ src/test/java/roomescape/controller/api/admin/AdminMemberApiControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api;
+package roomescape.controller.api.admin;
 
 import static org.hamcrest.Matchers.is;
 
@@ -6,11 +6,10 @@ import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 import roomescape.controller.BaseControllerTest;
 import roomescape.util.TokenGenerator;
 
-class MemberApiControllerTest extends BaseControllerTest {
+class AdminMemberApiControllerTest extends BaseControllerTest {
 
     @Override
     @BeforeEach
@@ -23,7 +22,7 @@ class MemberApiControllerTest extends BaseControllerTest {
     void selectMembers_Success() {
         RestAssured.given().log().all()
                 .cookie("token", TokenGenerator.makeAdminToken())
-                .when().get("/members")
+                .when().get("/admin/members")
                 .then().log().all()
                 .statusCode(200)
                 .body("members.size()", is(2));
